# contracts/openapi/core-api.v1.yaml
openapi: 3.0.3
info:
  title: FLIT Core API
  version: "1.0.0"
  description: >
    Contrato público de FLIT Core API (services/core-api). Cubre el carril de
    Seguridad/Autenticación (HU #10168 — login email/contraseña y emisión JWT 12h)
    y el módulo Administrador de Compañías (HU #10189 — index paginado y
    seguridad SuperAdmin).
servers:
  - url: https://api.flit.co
    description: Producción (vía Gateway YARP)
  - url: http://localhost:4002
    description: Local (Gateway)
...
paths:
...
  /api/v1/tramites:
    post:
      tags: [Trámites · Documentos]
      operationId: tramitesCreate
      summary: Crea un trámite y congela su snapshot documental (RF19)
      description: >
        Crea un trámite en estado draft y captura, en la misma transacción, un snapshot
        inmutable de la matriz documental resuelta vigente (precedencia Cliente > OT >
        Default). Si la generación del snapshot falla, toda la operación se revierte y no
        queda trámite creado (500). Requiere JWT autenticado (no SuperAdmin).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureInstanceRequest"
      responses:
        "201":
          description: Trámite creado con su snapshot documental.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcedureInstanceResponse"
        "401":
          description: No autenticado (sin token JWT).
        "404":
          description: El tipo de trámite o el cliente (tenant) no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Datos inválidos o número de referencia duplicado para el cliente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Falló la persistencia del trámite o del snapshot; operación revertida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/tramites/biometric-validations:
    get:
      operationId: listTenantBiometricValidations
      tags: [Tramites, Identidad]
      summary: Listar las validaciones de identidad del tenant (HU #10234, filtros HU #10347)
      description: >
        Vista transversal del submódulo "Validaciones de Identidad": lista TODAS las validaciones
        biométricas/de identidad del tenant (a través de todas sus instancias no eliminadas), las más
        recientes primero, junto con KPIs agregados por estado. No es por-instancia. La tabla se acota a
        un máximo de filas (vista de monitoreo); los KPIs se calculan aparte y son exactos. Con filtros
        activos, filas y KPIs reflejan el mismo subconjunto filtrado (antes del cap de 500 filas). El
        motivo de rechazo viaja SANITIZADO (sin PII), igual que en la vista por-instancia. Filtrar por
        estado=expirado incluye validaciones con estado persistido expirado y las que tienen el flag
        expired calculado (no aprobada y ExpiresAt &lt; now).
      security:
        - bearerAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: referenceNumber
          in: query
          required: false
          description: Coincidencia parcial (case-insensitive) sobre el número de referencia del trámite.
          schema: { type: string, maxLength: 100 }
        - name: modalidad
          in: query
          required: false
          description: Coincidencia parcial (case-insensitive) sobre la modalidad del trámite.
          schema: { type: string }
        - name: name
          in: query
          required: false
          description: Coincidencia parcial (case-insensitive) sobre el nombre de la persona validada.
          schema: { type: string, maxLength: 100 }
        - name: partyRole
          in: query
          required: false
          description: Filtro exacto por parte (comprador | vendedor). Los VALORES siguen en español.
          schema: { type: string, enum: [comprador, vendedor] }
        - name: documentType
          in: query
          required: false
          description: Filtro exacto por tipo de documento (p.ej. CC, CE).
          schema: { type: string, maxLength: 100 }
        - name: documentNumber
          in: query
          required: false
          description: Coincidencia parcial por número de documento (sin exponer PII en errores).
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          required: false
          description: >
            Filtro por estado. Valores (en español): enviado, en_proceso, aprobado, rechazado, expirado. El
            valor expirado incluye estado persistido y validaciones vencidas no aprobadas (flag expired).
          schema:
            type: string
            enum: [enviado, en_proceso, aprobado, rechazado, expirado, pendiente_envio, error_envio]
        - name: provider
          in: query
          required: false
          description: Filtro exacto por proveedor (mock | kyverum).
          schema: { type: string, enum: [mock, kyverum] }
        - name: scoreMin
          in: query
          required: false
          description: Score mínimo inclusive (solo filas con score informado).
          schema: { type: integer, minimum: 0 }
        - name: scoreMax
          in: query
          required: false
          description: Score máximo inclusive (solo filas con score informado).
          schema: { type: integer, minimum: 0 }
        - name: createdFrom
          in: query
          required: false
          description: Fecha/hora mínima inclusive de creación (ISO-8601).
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          required: false
          description: Fecha/hora máxima inclusive de creación (ISO-8601).
          schema: { type: string, format: date-time }
        - name: motivoRechazo
          in: query
          required: false
          description: >
            Coincidencia parcial case-insensitive en el motivo de rechazo sanitizado (solo filas
            rechazadas; busca en Detalle y ProviderPayload).
          schema: { type: string, maxLength: 100 }
        - name: vigenciaEstado
          in: query
          required: false
          description: >
            Filtro por estado de vigencia de la identidad APROBADA: vigente | por_vencer (≤ 7 días) | vencida.
          schema: { type: string, enum: [vigente, por_vencer, vencida] }
        - name: expiraDesde
          in: query
          required: false
          description: Fin de vigencia (validUntil) mínimo inclusive, ISO-8601 (HU #10350).
          schema: { type: string, format: date-time }
        - name: expiraHasta
          in: query
          required: false
          description: Fin de vigencia (validUntil) máximo inclusive, ISO-8601 (HU #10350).
          schema: { type: string, format: date-time }
        - name: venceEnDias
          in: query
          required: false
          description: >
            "Vence en ≤ N días": identidades vigentes cuya vigencia se agota dentro de N días calendario.
          schema: { type: integer, minimum: 0 }
        - name: page
          in: query
          required: false
          description: Página (1-based). Por defecto 1; valores < 1 se normalizan a 1.
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          description: Filas por página. El cliente elige de 10 en 10 hasta 50; fuera de rango se acota.
          schema: { type: integer, minimum: 10, maximum: 50, default: 20 }
      responses:
        "200":
          description: Validaciones del tenant + KPIs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantBiometricValidationsResponse"
        "400":
          description: >
            Falta el header X-Tenant-Id, parámetros de filtro inválidos (estado/provider/parte fuera de
            catálogo, scoreMin &gt; scoreMax, createdFrom &gt; createdTo).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tramites/identity-validation/stuck:
    get:
      operationId: listStuckIdentityValidations
      tags: [Tramites, Identidad]
      summary: Eventos de validación de identidad atascados (dead-letter) (HU #10349 fase 2)
      description: >
        Observabilidad operativa del encadenamiento async de identidad (patrón outbox). Lista los eventos
        `completed` del tenant que quedaron ATASCADOS: pendientes (no despachados) que agotaron los
        reintentos del worker (`attempts >= maxDeliveryAttempts`) y, por lo tanto, NO se reintentan más.
        Útil para detectar trámites cuyo auto-flujo (firma/FUR) quedó trabado. Recuperación operativa:
        reencolar poniendo `attempts = 0` en la fila (el worker la retomará). Sin PII (no expone el payload).
      security:
        - bearerAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Eventos atascados + total.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StuckIdentityValidationsResponse"
        "400":
          description: Falta el header X-Tenant-Id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tramites/identity-validation/stuck/{id}/requeue:
    post:
      operationId: requeueStuckIdentityValidation
      tags: [Tramites, Identidad]
      summary: Reencolar ("desatascar") un evento de identidad atascado (HU #10349 fase 2)
      description: >
        Reinicia los intentos de un evento ATASCADO (dead-letter) para que el worker lo vuelva a procesar
        en el próximo ciclo. Idempotente. 404 si no hay un evento atascado con ese id para el tenant.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: Id del evento en la outbox (identity_validation_outbox.id).
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Evento reencolado; el worker lo retomará en el próximo ciclo.
          content:
            application/json:
              schema:
                type: object
                required: [requeued]
                properties:
                  requeued: { type: boolean, example: true }
        "400":
          description: Falta el header X-Tenant-Id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No hay un evento atascado con ese id para el tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/tramites/identity-validation/stuck/requeue-all:
    post:
      operationId: requeueAllStuckIdentityValidations
      tags: [Tramites, Identidad]
      summary: Reencolar TODOS los eventos de identidad atascados del tenant (HU #10349 fase 2)
      description: >
        Reinicia los intentos de todos los eventos ATASCADOS (dead-letter) del tenant de una vez para que
        el worker los vuelva a procesar. Idempotente. Devuelve cuántos se reencolaron.
      security:
        - bearerAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Cantidad de eventos reencolados.
          content:
            application/json:
              schema:
                type: object
                required: [requeued]
                properties:
                  requeued: { type: integer, example: 3 }
        "400":
          description: Falta el header X-Tenant-Id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/tramites/instances/{id}/identity/ensure:
    post:
      operationId: ensureProcedureInstanceIdentity
      tags: [Tramites, Identidad]
      summary: Asegurar la identidad vigente de una parte (HU #10350)
      description: >
        Al guardar una parte (comprador/vendedor) el sistema asegura su validación de identidad sin
        esperar un clic manual: si la persona (tipoDoc+documento) tiene una validación APROBADA y VIGENTE
        (≤30 días calendario desde la aprobación; el día de aprobación es el día 1 y vence el día 31) en
        cualquier trámite del tenant, la reutiliza clonándola a este trámite (hereda la fecha de
        aprobación → conserva el mismo vencimiento). Si no hay vigente, responde `requiere_validacion`
        para que el cliente sea enviado a validar automáticamente. Idempotente.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parte]
              properties:
                parte:
                  type: string
                  enum: [comprador, vendedor]
      responses:
        "200":
          description: Desenlace de la verificación de identidad de la parte.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnsureIdentityResult"
        "400":
          description: parte inválida o falta X-Tenant-Id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Instancia no encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CreateProcedureInstanceRequest:
      type: object
      properties:
        procedureTypeId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        transitOfficeId:
          type: string
          format: uuid
          nullable: true
          description: Organismo de tránsito (nivel OT) — opcional, afecta la precedencia.
        referenceNumber:
          type: string
          maxLength: 30
          description: Número de referencia del trámite, único por cliente.
        createdByUserId:
          type: string
          format: uuid
          nullable: true
          description: Usuario creador; si se omite, se resuelve del JWT (sub).
      required: [procedureTypeId, tenantId, referenceNumber]
    ProcedureInstanceResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        status: { type: string, example: draft }
        snapshotAt: { type: string, format: date-time }
        documentCount: { type: integer, description: Documentos congelados en el snapshot. }
      required: [id, referenceNumber, status, snapshotAt, documentCount]
...
    BiometricValidationStatsDto:
      type: object
      required: [total, aprobadas, enProceso, rechazadas, expiradas]
      description: KPIs agregados del submódulo de Validaciones (exactos, sin el cap de filas).
      properties:
        total: { type: integer }
        aprobadas: { type: integer }
        enProceso:
          type: integer
          description: Validaciones en curso (enviado + en_proceso).
        rechazadas: { type: integer }
        expiradas: { type: integer }
    TenantBiometricValidationsResponse:
      type: object
      required: [validations, stats, page, pageSize, total]
      properties:
        validations:
          type: array
          description: Filas de la PÁGINA pedida (no todo el conjunto).
          items:
            $ref: "#/components/schemas/TenantBiometricValidationDto"
        stats:
          $ref: "#/components/schemas/BiometricValidationStatsDto"
        page:
          type: integer
          description: Página devuelta (1-based).
        pageSize:
          type: integer
          description: Filas por página efectivas (acotadas a [10, 50]).
        total:
          type: integer
          description: Total de filas del conjunto filtrado completo (para calcular el nº de páginas).
...